requirements.txt
arxiv>=2.1.0
anthropic>=0.40.0
pyyaml>=6.0
markdown>=3.5
fpdf2>=2.8.0
flask>=3.0.0
